change name of a lot of classes, clean structure,working branch
