--- src/model/Loginable.java
@@ -1,7 +1,0 @@
-package model;
-
-public interface Loginable {
-
-    public void login();
-
-}
