--- src/model/Searcher.java
@@ -1,8 +1,0 @@
-package model;
-
-import parsers.MovieInfo;
-
-public interface Searcher {
-
-    public MovieInfo search(String movieTitle);
-}
