+++ src/model/WebDriverHandler.java
@@ -1,6 +1,32 @@
 package model;
 
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+
+import java.util.HashMap;
+
 public class WebDriverHandler {
 
-    
+    //protected String rootUrl; //Maybe remove or change
+    protected ChromeDriver browser;
+    protected ChromeOptions options;
+
+    public WebDriverHandler() {
+        System.setProperty("webdriver.chrome.driver", "chromedriver.exe");
+        this.options = new ChromeOptions();
+        options.setHeadless(true);
+
+        //this.browser = new ChromeDriver(options);
+        //this.cookieHandler = new MyCookieHandler(options, fileName);
+        //this.account = new UserAccount();
+        //this.rootUrl = rootUrl;
+
+        //set preferences to not load images and to use disk cache
+        //--------------------------------------------------
+        HashMap<String, Object> prefs = new HashMap<String, Object>();
+        prefs.put("profile.managed_default_content_settings.images", 2);
+        prefs.put("disk-cache-size", 4096);
+        options.setExperimentalOption("prefs", prefs);
+        //--------------------------------------------------
+    }
 }
