--- src/parsers/Parsable.java
@@ -1,5 +1,0 @@
-package parsers;
-
-public interface Parsable {
-    public void parse(String movieTitle);
-}
